Playing with SPI. Fails after one transaction

+++ firmware/include/peripherals/spi.h
@@ -5,7 +5,7 @@
 #include <stdint.h>
 
 int8_t SPI_init(uint32_t baud, uint8_t clock_mode);
-int8_t SPI_transfer(uint8_t length, uint8_t *data_out, uint8_t *data_in);
+int8_t SPI_transfer(uint16_t length, uint8_t *data_out, uint8_t *data_in);
 int8_t SPI_close();
 
 #endif
